working screen fading mechanic
