+++ src/Context/ScreensContext.jsx
@@ -3,10 +3,13 @@ import { useState, createContext, useContext } from 'react';
 export const ScreensContext = createContext();
 
 export function ScreensProvider(props) {
-	const [currentScreenTitle, setCurrentScreenByTitle] = useState('intro');
+	const [currentScreenTitle, setCurrentScreenByTitle] = useState('welcome');
+	const [nextScreenTitle, setNextScreenByTitle] = useState(null);
 
 	return (
-		<ScreensContext.Provider value={{ currentScreenTitle, setCurrentScreenByTitle }}>
+		<ScreensContext.Provider
+			value={{ currentScreenTitle, setCurrentScreenByTitle, nextScreenTitle, setNextScreenByTitle }}
+		>
 			{props.children}
 		</ScreensContext.Provider>
 	);
