+++ src/Screens/WelcomeScreen/HiddenMessage.jsx
@@ -2,8 +2,8 @@ import { texts } from 'Texts/texts';
 import { Text } from 'Texts/Text';
 import { HiddenMessageTextContainer } from './styles';
 
-export const HiddenMessageScreen = () => {
-	const hiddenMessageText = texts.introScreen.hiddenMessage;
+export const HiddenMessage = () => {
+	const hiddenMessageText = texts.welcomeScreen.hiddenMessage;
 	return (
 		<HiddenMessageTextContainer>
 			<Text>{hiddenMessageText}</Text>
