+++ src/Screens/screensLib.js
@@ -1,7 +1,9 @@
 import { IntroScreen } from './IntroScreen/IntroScreen';
 import { MoneyScreen } from './MoneyScreen/MoneyScreen';
+import { WelcomeScreen } from './WelcomeScreen/WelcomeScreen';
 
 export const screensLib = {
+	welcome: WelcomeScreen,
 	intro: IntroScreen,
 	money: MoneyScreen,
 };
